Update README.md
This program performs CNN classification on the Cifr10 dataset with the best accuracy of 78%. I used preprocessing image transformation from torch libraries to enhance the performance accuracy of the data. Using, RamdomCrop, RandomHorizontralFlip() and normalization transformations on training data. 

for the CNN I used a 2 layer convolutional network and a relu activation function in the forward propagation.